电话转接增加houseId和houseType字段

+++ Article/Bubble/Modules/Bizs/HorseDetailPage/PageVM/FollowUpViewModel.swift
@@ -84,7 +84,7 @@ class FollowUpViewModel {
 
     func followThisItem(
         isFollowUpOrCancel: Bool,
-        houseId: Int,
+        houseId: Int64,
         statusBehavior: BehaviorRelay<Bool>) {
         if EnvContext.shared.client.accountConfig.userInfo.value == nil {
             displayLogin(onLoginSuccess: { [unowned self] in
